feat(SpokePool): Modify existing deposit() function to emit USSFundsDeposited event and add new depositUSS function
Fixes #ACX-1684 and #ACX-1685

## Modifications to existing deposit function: deposit():
- At upgrade: backwards compatible deposit() function will emit the new event. The deposit2() function (with new params) will emit the new event as well. There will be no deposit function post upgrade that emits the old event. The idea behind modifying the existing function is that post-upgrade, no more “old deposits” are emitted.
- Instead of the depositor using the relayerFeePct parameter to set the relayer’s portion of the fee, depositor uses the relayerFeePct to mark a “total fee” percentage. The recipient therefore knows that they will receive `amount * (1 - relayerFeePct)` tokens. The funds deposited event emits outputTokenAmount equal to this amount * (1 - relayerFeePct).
- outputToken: new parameter, defaults to some default token address like 0x0 to indicate that the same input and output token should be used
- Should hardcode committedFiller to 0x0
- Set expiryTimestamp to MAX_UINT. We never want to process deposit refunds for “old deposits”. We should add a rule in the dataworker that deposit refunds only count for type 2 deposits.

## Provides new public function deposit2() for depositors:
- expiryTimestamp: integer
    - This is going to be relative to the destination chain. Don’t allow this to be more than ~3 bundles into the future, so 12 hours maybe. This way the dataworker/relayer doesn’t have to maintain more a lookback longer than this.
- Emits FundsDeposited2 event

+++ contracts/test/MockSpokePool.sol
@@ -10,7 +10,7 @@ import "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
  */
 contract MockSpokePool is SpokePool, OwnableUpgradeable {
     uint256 private chainId_;
-    uint256 private currentTime;
+    uint32 private currentTime;
 
     function initialize(
         uint32 _initialDepositId,
@@ -20,14 +20,14 @@ contract MockSpokePool is SpokePool, OwnableUpgradeable {
     ) public initializer {
         __Ownable_init();
         __SpokePool_init(_initialDepositId, _crossDomainAdmin, _hubPool, _wethAddress);
-        currentTime = block.timestamp; // solhint-disable-line not-rely-on-time
+        currentTime = uint32(block.timestamp); // solhint-disable-line not-rely-on-time
     }
 
-    function setCurrentTime(uint256 time) external {
+    function setCurrentTime(uint32 time) external {
         currentTime = time;
     }
 
-    function getCurrentTime() public view override returns (uint256) {
+    function getCurrentTime() public view override returns (uint32) {
         return currentTime;
     }
